requirements.txt
# Core Dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
PyYAML>=6.0
typing-extensions>=4.0.0

# OpenAI API
openai>=1.0.0

# PDF Processing
PyPDF2>=3.0.0

# Data Validation
pydantic>=1.8.0

# Development & Testing
pytest>=6.2.5
pytest-asyncio>=0.15.1
httpx>=0.23.0

# Code Quality
black>=21.12b0
isort>=5.10.1
mypy>=0.930
flake8>=4.0.1

# Documentation
mkdocs>=1.2.3
mkdocs-material>=8.1.0
